gralloc: Raise sNumInts field to 24
While 20 seems to fit Exynos5430 camera blobs (libexynoscamera.so),
it isn't matching the values we get when disassembling 5433 blobs
(from a Galaxy Note 4 in this case).

Credit goes to Andreas Schneider for making me understand the ARM
assembler code portions and analyzing the pseudo-code containing
the if-conditions which were leading to the invalid gralloc handles
we have been seeing.

Change-Id: Iadce01be16968ffd2c5de12df52a1f05313ae07b